Move functions from LocalNode to RemoteNode that only use the latter.
